reorganize product readers into several smaller modules
