+++ src/nisarqa/products/__init__.py
@@ -1,3 +1,2 @@
 # keep individual products in their own namespace
 from . import caltools, gcov, gslc, igram, offsets, rslc
-from .product_reader import *
